fixed thresholder

+++ core/E_DicomSeries.h
@@ -18,7 +18,7 @@ class E_DicomSeries{
     ~E_DicomSeries();
 
     
-    typedef itk::Image<float, 3> ImageType;
+    typedef itk::Image<short, 3> ImageType;
     typedef itk::ImageSeriesReader<ImageType> DicomReader;
     typedef itk::GDCMImageIO ImageIOType;
