+++ core/E_ImageSeries.h
@@ -13,7 +13,7 @@ class E_ImageSeries{
     ~E_ImageSeries();
 
     
-    typedef itk::Image<float, 3> ImageType;
+    typedef itk::Image<short, 3> ImageType;
     typedef itk::ImageSeriesReader<ImageType> DicomReader;
     typedef itk::GDCMImageIO ImageIOType;
